Create helper.js file for controllers and move there shared functions
