+++ backend/controllers/application.controller.js
@@ -18,60 +18,16 @@ import Job from "../models/job.model.js";
 import User from "../models/user/user.model.js";
 import CodeAssessment from "../models/codeAssessment.js";
 import CodeSubmission from "../models/codeSubmission.js";
-
-/**
- * Creates a standardized API response object
- * @param {boolean} success - Indicates if the operation was successful
- * @param {string} message - Response message
- * @param {Object} [data] - Response data (optional)
- * @returns {Object} Standardized response object
- */
-export const createResponse = (success, message, data = null) => ({
-    success,
-    message,
-    ...(data && { data }),
-});
-
-/**
- * Handles errors and sends appropriate response
- * @param {Object} res - Express response object
- * @param {Error} error - Error object
- * @param {string} [defaultMessage] - Default error message
- * @returns {Object} Error response
- */
-export const handleError = (res, error, defaultMessage = "Server error") => {
-    const statusCode = error.status || 500;
-    return res.status(statusCode).json(createResponse(false, error.message || defaultMessage));
-};
-
-/**
- * Validates application status
- * @param {string} status - Status to validate
- * @returns {boolean} Whether status is valid
- */
-export const isValidStatus = (status) => {
-    const validStatuses = ["Applying", "Applied", "In Review", "Shortlisted", "Rejected", "Accepted"];
-    return validStatuses.includes(status);
-};
-
-/**
- * Verifies job ownership
- * @param {string} jobId - Job ID to verify
- * @param {string} userId - User ID to check against
- * @returns {Promise<boolean>} Whether user owns the job
- */
-export const verifyJobOwnership = async (jobId, userId) => {
-    try {
-        const job = await Job.findById(jobId);
-        if(job){
-            return job.postedBy.toString() == userId;
-        } else {
-            throw new Error("Job not found");
-        }
-    } catch (err){
-        return false;
-    }
-};
+import { 
+    createResponse, 
+    handleError, 
+    formatAnswers, 
+    isValidStatus, 
+    verifyJobOwnership,
+    getTotalStatus,
+    getJobsByEmployerHelper,
+    getLineGraphData 
+} from "./helpers.js";
 
 /**
  * Application controller handling all application-related operations
@@ -365,43 +321,6 @@ export const applicationController = {
             return await handleStatusProgression(res, app, user);
         } catch (error) {
             return handleError(res, error, "Error updating application status");
-
-//             if (!app) {
-//                 return res.status(404).json(createResponse(false, "Application not found"));
-//             }
-//             console.log("toReject: ", toReject);
-//             if(toReject){
-//                 if(app.status === "Accepted"){
-//                     return res.status(400).json(createResponse(false, "Cannot reject an accepted application"));
-//                 }
-//                 app.status = "Rejected";
-//                 await app.save();
-//                 return res.json(createResponse(true, "Application rejected", app));
-//             }
-//             if(app.status.trim() === "Code Challenge" && user.role.trim() === "employer"){
-//                 return res.status(400).json(createResponse(false, "Cannot update status"));
-//             }
-//             if(app.status !== "Code Challenge" && user.role === "applicant"){
-//                 return res.status(400).json(createResponse(false, "Cannot update status"));
-//             }
-
-//             const hasCodeAssessment = app.job.assessments.length > 0;
-//             console.log("hasCodeAssessment: ", hasCodeAssessment);
-//             const statuses = ['Applied', 'Shortlisted', 'Code Challenge', 'In Review', 'Accepted'];
-//             const currentIndex = statuses.indexOf(app.status);
-//             if (currentIndex === -1 || currentIndex === statuses.length - 1) {
-//                 return res.status(400).json(createResponse(false, "No further status available"));
-//             }
-//             app.status = statuses[currentIndex + 1];
-//             console.log("app.status (chnaged): ", app.status);
-//             if(app.status === 'Code Challenge' && !hasCodeAssessment) {
-//                 app.status = statuses[currentIndex + 2];
-//             }
-//             await app.save();
-//             res.json(createResponse(true, "Application status updated", app));
-//         } catch (err) {
-//             console.error("Error updating application status:", err);
-//             res.status(500).json(createResponse(false, err.message));
         }
     },
 
@@ -502,175 +421,3 @@ export const applicationController = {
         }
     }
 };
-
-/**
- * Formats answers for storage
- * @param {Array} answers - Raw answers array
- * @returns {Array} Formatted answers with ObjectId
- */
-export const formatAnswers = (answers) => 
-    answers.map(answer => ({
-        questionId: new mongoose.Types.ObjectId(answer.questionId),
-        answerText: answer.answerText,
-    }));
-
-/**
- * Handles application rejection
- * @param {Object} res - Express response object
- * @param {Object} application - Application document
- * @returns {Promise<void>}
- */
-export const handleRejection = async (res, application) => {
-    if (application.status === "accepted") {
-        return res.status(400).json(createResponse(false, "Cannot reject an accepted application"));
-    }
-    application.status = "rejected";
-    await application.save();
-    return res.json(createResponse(true, "Application rejected", application));
-};
-
-async function getJobsByEmployerHelper(uid) {
-    try {
-        const employer = await User.findOne({ uid  });
-        if (!employer) {
-            throw new Error("Employer not found");
-        }
-        const jobs = await Job.find({ postedBy: employer._id });
-        return jobs;
-    
-    } catch (error) {
-        console.error("Error fetching employer jobs:", error);
-        throw new Error(error.message);
-    }
-    
-}
-
-/**
- * Handles application status progression
- * @param {Object} res - Express response object
- * @param {Object} application - Application document
- * @param {Object} user - User document
- * @returns {Promise<void>}
- */
-export const handleStatusProgression = async (res, application, user) => {
-    const statuses = ['Applied', 'Shortlisted', 'Code Challenge', 'In Review', 'Accepted'];
-    const currentIndex = statuses.indexOf(application.status);
-
-    if (user.role === "employer" && application.status === "Code Challenge") {
-        return res.status(400).json(createResponse(false, "Cannot update status"));
-    }
-    if (user.role === "applicant" && application.status !== "Code Challenge") {
-        return res.status(400).json(createResponse(false, "Cannot update status"));
-    }
-    if (currentIndex === -1 || currentIndex === statuses.length - 1) {
-        return res.status(400).json(createResponse(false, "No further status available"));
-    }
-
-    application.status = statuses[currentIndex + 1];
-    if (application.status === 'Code Challenge' && !application.job.assessments.length) {
-        application.status = statuses[currentIndex + 2];
-    }
-    await application.save();
-    return res.json(createResponse(true, "Application status updated", application));
-};
-
-/**
- * Retrieves jobs by employer
- * @param {string} uid - User ID
- * @returns {Promise<Array>} Array of jobs
- */
-export const getJobsByEmployer = async (uid) => {
-    const employer = await User.findOne({ uid });
-    if (!employer) throw new Error("Employer not found");
-    return Job.find({ postedBy: employer._id });
-};
-
-    async function getLineGraphData(jobs) {
-    try {
-        const jobIds = jobs.map(job => job._id);
-        const lineGraphData = await Application.aggregate([
-            { 
-                $match: { 
-                    job: { $in: jobIds }, 
-                    status: { $ne: "Applying" } // Exclude applications in the "applying" state
-                } 
-            },
-            { 
-                $group: { 
-                    _id: { 
-                        jobId: "$job", // Include jobId in the grouping
-                        date: { $dateToString: { format: "%d-%m-%Y", date: "$submittedAt" } } // Group by date
-                    }, 
-                    count: { $sum: 1 } // Count applications for each date
-                } 
-            },
-            { $sort: { "_id.date": 1 } }, // Sort by date in ascending order
-            { 
-                $project: { 
-                    jobId: "$_id.jobId", // Include jobId in the final output
-                    date: "$_id.date",
-                    count: 1,
-                    _id: 0 // Remove the _id field
-                } 
-            }
-        ]);
-        return lineGraphData;
-    } catch (error) {
-        console.error("Error fetching line graph data:", error);
-        throw new Error("Failed to fetch line graph data");
-    }
-}
-
-/**
- * Builds dashboard data object
- * @param {Object} employer - Employer document
- * @param {Array} jobs - Array of job documents
- * @returns {Promise<Object>} Dashboard data
- */
-export const buildDashboardData = async (employer, jobs) => {
-    const totalStatus = await getTotalStatus(jobs);
-    const lineGraphData = await getLineGraphData(jobs);
-    return {
-        totalJobs: jobs.length,
-        totalStatus,
-        companyName: employer.companyName || "",
-        jobs,
-        lineGraphData,
-    };
-};
-
-/**
- * Gets total status counts for jobs
- * @param {Array} jobs - Array of job documents
- * @returns {Promise<Array>} Status aggregation
- */
-export const getTotalStatus = async (jobs) => {
-    const jobIds = jobs.map(job => job._id);
-    return Application.aggregate([
-        { $match: { job: { $in: jobIds } } },
-        { $group: { _id: "$status", count: { $sum: 1 } } },
-    ]);
-};
-
-// /**
-//  * Gets line graph data for job applications
-//  * @param {Array} jobs - Array of job documents
-//  * @returns {Promise<Array>} Line graph data
-//  */
-// export const getLineGraphData = async (jobs) => {
-//     const jobIds = jobs.map(job => job._id);
-//     return Application.aggregate([
-//         { $match: { job: { $in: jobIds }, status: { $ne: "applying" } } },
-//         { 
-//             $group: { 
-//                 _id: { 
-//                     jobId: "$job",
-//                     date: { $dateToString: { format: "%Y-%m-%d", date: "$submittedAt" } }
-//                 }, 
-//                 count: { $sum: 1 }
-//             } 
-//         },
-//         { $sort: { "_id.date": 1 } },
-//         { $project: { jobId: "$_id.jobId", date: "$_id.date", count: 1, _id: 0 } }
-//     ]);
-// };
