+++ backend/controllers/user.controller.js
@@ -1,19 +1,7 @@
 import User from "../models/user/user.model.js";
 import { JobSeeker } from "../models/user/jobSeeker.model.js";
 import { Employer } from "../models/user/Employer.model.js";
-
-/**
- * Creates a standardized response object
- * @param {boolean} success - Indicates operation success
- * @param {string} message - Response message
- * @param {Object} [data] - Response data (optional)
- * @returns {Object} Standardized response
- */
-const createResponse = (success, message, data = null) => ({
-    success,
-    message,
-    ...(data && { data }),
-});
+import { createResponse } from "./helpers.js";
 
 /**
  * Validates user authentication
